Migrated the relationship handling to an external object. Improved order of operations when executing command buffers.
