+++ psl/src.txt
@@ -20,6 +20,7 @@ profiling/profiler
 
 ecs/state
 ecs/details/component_container
+ecs/details/entity_relationship_handler
 ecs/command_buffer
 
 async/scheduler
